Move automation code to video.calling package. Split by feature.
