+++ TelepresenceBot/mobile/src/main/AndroidManifest.xml
@@ -17,7 +17,7 @@
     tools:ignore="AllowBackup,GoogleAppIndexingWarning">
 
     <service
-      android:name="com.novoda.tpbot.automation.HangoutJoinerAutomationService"
+      android:name=".bot.video.calling.HangoutJoinerAutomationService"
       android:permission="android.permission.BIND_ACCESSIBILITY_SERVICE">
       <intent-filter>
         <action android:name="android.accessibilityservice.AccessibilityService" />
